requirements.txt
certifi==2017.4.17
chardet==3.0.3
click==6.7
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
pyslack==0.3.0
requests==2.17.3
urllib3==1.21.1
Werkzeug==0.12.2
